Working version of tracker
Changed Arduino and python code.
Synchronisation works as follows at the moment:
Teensy 3.5 gets as input the frame clock - whenever it registers a HIGH and the timer is > specified interval (e.g. 20 ms for 50 Hz) it then writes the current rotary encoder value via serial. While it is writing to serial, it also sends a HIGH on one of the National Instruments connected pins (the 6363 card in this case) and terminates with a LOW on that line. That way serial and NI registered events are in sync.
Everything runs in the exact same wavesurfer protocol as before which has only been modified to record one more digital input line (the wheel HIGH / LOW during serial write).
The python script receives these values and saves them to a .csv with current system timestamp.